Issue #26
It allows to pass views contextual filters as path arguments. This is, you could filter posts by tid (which needs to be done through contextual filters - exposed filters only allow to filter by term name) by adding a contextual filter to the view. Then you can just add the tid to the endpoint like this:

```
/api/v2/views/posts/page/12?page=1
```
This will show the posts which have the term id 12.

This is for the views v2 end point only.